requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
mysql-connector==2.2.9
paho-mqtt==1.4.0
phpserialize==1.3
PyYAML==5.1
requests==2.21.0
urllib3==1.24.2
